requirements.txt
python-dotenv~=1.0.1
langchain_google_genai~=2.0.7
langchain_core~=0.3.28
python-dotenv~=1.0.1
streamlit~=1.41.1
pillow~=11.0.0